# api-gateway/swagger.yaml
openapi: 3.0.3
info:
  title: CommerceAI Pro API
  description: |
    API Gateway pour la plateforme CommerceAI Pro - Intelligence Artificielle pour E-commerce
    
    Cette API fournit un accès complet aux fonctionnalités de la plateforme :
    - Authentification et gestion des utilisateurs
    - Gestion des organisations et rôles
    - Support client et chat en temps réel
    - Analytics et métriques
    - Intégrations avec les plateformes e-commerce
    - Agents IA spécialisés
    
    ## Authentification
    
    L'API utilise l'authentification JWT. Incluez le token dans l'en-tête Authorization :
    ```
    Authorization: Bearer <votre-token-jwt>
    ```
    
    ## Codes de statut
    
    - `200` - Succès
    - `201` - Créé avec succès
    - `400` - Erreur de validation
    - `401` - Non authentifié
    - `403` - Accès refusé
    - `404` - Ressource non trouvée
    - `429` - Limite de taux dépassée
    - `500` - Erreur serveur
    
  version: 1.0.0
  contact:
    name: Support CommerceAI Pro
    email: support@commerceai-pro.com
    url: https://commerceai-pro.com
  license:
    name: Propriétaire
    url: https://commerceai-pro.com/license
servers:
  - url: https://api.commerceai-pro.com
    description: Production
  - url: https://staging-api.commerceai-pro.com
    description: Staging
  - url: http://localhost:4000
    description: Développement local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Erreur lors de l'opération"
        error:
          type: string
          example: "Détails de l'erreur"
        timestamp:
          type: string
          format: date-time
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Opération réussie"
        data:
          type: object
        timestamp:
          type: string
          format: date-time
...
    Ticket:
      type: object
      properties:
        id:
          type: string
          example: "ticket_123"
        number:
          type: string
          example: "TICK-2024-001"
        title:
          type: string
          example: "Problème de connexion"
        description:
          type: string
          example: "Je n'arrive pas à me connecter à mon compte"
        category:
          type: string
          enum: [technical, billing, general, feature_request]
          example: "technical"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          example: "medium"
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
          example: "open"
        userId:
          type: string
          example: "user_123"
        assignedTo:
          type: string
          example: "agent_456"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
paths:
...
  /api/support/tickets:
    get:
      tags:
        - Support
      summary: Liste des tickets de support
      description: Récupérer tous les tickets de l'utilisateur connecté
      parameters:
        - name: status
          in: query
          description: Filtrer par statut
          schema:
            type: string
            enum: [open, in_progress, resolved, closed]
        - name: category
          in: query
          description: Filtrer par catégorie
          schema:
            type: string
            enum: [technical, billing, general, feature_request]
        - name: page
          in: query
          description: Numéro de page
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Nombre d'éléments par page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Liste des tickets
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          tickets:
                            type: array
                            items:
                              $ref: '#/components/schemas/Ticket'
                          pagination:
                            type: object
                            properties:
                              page:
                                type: integer
                              limit:
                                type: integer
                              total:
                                type: integer
                              pages:
                                type: integer
    
    post:
      tags:
        - Support
      summary: Créer un nouveau ticket
      description: Créer un ticket de support
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - category
              properties:
                title:
                  type: string
                  example: "Problème de connexion"
                description:
                  type: string
                  example: "Je n'arrive pas à me connecter à mon compte"
                category:
                  type: string
                  enum: [technical, billing, general, feature_request]
                  example: "technical"
                priority:
                  type: string
                  enum: [low, medium, high, urgent]
                  example: "medium"
      responses:
        '201':
          description: Ticket créé avec succès
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Ticket'
  /api/support/tickets/{ticketId}:
    get:
      tags:
        - Support
      summary: Détails d'un ticket
      description: Récupérer les détails d'un ticket spécifique
      parameters:
        - name: ticketId
          in: path
          required: true
          description: ID du ticket
          schema:
            type: string
      responses:
        '200':
          description: Détails du ticket
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Ticket'
        '404':
          description: Ticket non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  # Chat Support
...